+++ CHANGES.md
@@ -10,6 +10,7 @@
 - Update various docstrings.
 - Optimize performance for`css_first` in lexbor backend
 - Fix segfaults when accessing attributes. Resolves #135.
+- Add new `.clone` method to lexbor backend. Resolve #117.
 
 ## Version 0.3.34